Improve error handling in process-compliance.sh
The retries on file copies TO our DOC1 server (added in
https://github.com/companieshouse/chips-weblogic-batch/pull/70) would
not generate failure alerts in most cases, as the script hit the
existing check for files returned FROM the DOC1 server first and
terminated on that.

Adjusting the error handling to make cause of errors clearer; also (after
team discussion) allowing onward processing to continue if there is at
least one file to print.

Changes made:
- Added check on status of copies to DOC1 server:
	- if all copies failed, terminate with an alert
	- if at least one copy succeeded, continue processing (with an
	  alarm if any failures)
- Switched threshold of existing error handling on checking for files
  returned from DOC1 server (currently terminates if >1 failure), to
  use the same principle:
	- if no files found in output directory, terminate with an alert
	- if at least one file found in output directory, continue
	  processing (with an alarm if any files missing)
- Adjusted checks for files returned from DOC1 server to make logging &
  alerting more accurate
- Stopped letter counts file from being regenerated when running in
  interactive repair mode using IGNORE_AFP_COUNT flag, e.g. to re-send
  individual missed files to gateway

When the compliance process is run end-to-end, the expected number of
AFP-format files returned from the DOC1 server will be based on the
number successfully submitted. When the interactive repair script is run
from after the DOC1 phase, this number will not be available, so the
script will expect the total number eligible for AFP processing (as at
present).